requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bitarray==3.8.0
blinker==1.9.0
cachetools==7.0.3
certifi==2026.2.25
charset-normalizer==3.4.5
ckzg==2.1.7
click==8.3.1
colorama==0.4.6
cytoolz==1.1.0
Deprecated==1.3.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.135.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hexbytes==1.3.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
limits==5.8.0
MarkupSafe==3.0.3
multidict==6.7.1
narwhals==2.17.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
parsimonious==0.10.0
pillow==12.1.1
plotly==6.6.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.1
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
pyunormalize==17.0.0
pywin32==311
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rlp==4.1.0
rpds-py==0.30.0
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
starlette==0.52.1
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
web3==7.14.1
websockets==15.0.1
wrapt==2.1.2
yarl==1.23.0
